requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
pydantic==2.4.2
requests==2.31.0
beautifulsoup4==4.12.2
pillow==10.1.0
colorthief==0.2.1
python-multipart==0.0.6
pytest==7.4.3
httpx==0.25.1
aiohttp==3.8.6
playwright==1.39.0
redis==5.0.1
dotenv